pyproject.toml
[project]
name = "whisper-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "openai-whisper>=20250625",
    "python-dotenv>=1.2.2",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "pandas>=2.2.0",
]

[project.scripts]
whisper-app = "whisper_app.cli:app"

